# Optimize subdivide 3