Before
--------
```
faces     | vertices  | halfedges | edges     | verts/s   | time (ms)
----------+-----------+-----------+-----------+-----------+-----------
24        | 26        | 96        | 48        | 389678    | 0.05     
96        | 98        | 384       | 192       | 474868    | 0.15     
384       | 386       | 1536      | 768       | 670404    | 0.43     
1536      | 1538      | 6144      | 3072      | 647458    | 1.78     
6144      | 6146      | 24576     | 12288     | 671209    | 6.87     
24576     | 24578     | 98304     | 49152     | 561476    | 32.83    
98304     | 98306     | 393216    | 196608    | 535278    | 137.74   
393216    | 393218    | 1572864   | 786432    | 507164    | 581.49   
1572864   | 1572866   | 6291456   | 3145728   | 485607    | 2429.22  
Vertices per second, min: 389678, max: 671209, avg: 549238
```

After
------
```
faces     | vertices  | halfedges | edges     | verts/s   | time (ms)
----------+-----------+-----------+-----------+-----------+-----------
24        | 26        | 96        | 48        | 707575    | 0.03     
96        | 98        | 384       | 192       | 919341    | 0.08     
384       | 386       | 1536      | 768       | 1031016   | 0.28     
1536      | 1538      | 6144      | 3072      | 1032396   | 1.12     
6144      | 6146      | 24576     | 12288     | 1034798   | 4.45     
24576     | 24578     | 98304     | 49152     | 892153    | 20.66    
98304     | 98306     | 393216    | 196608    | 804561    | 91.64    
393216    | 393218    | 1572864   | 786432    | 782851    | 376.72   
1572864   | 1572866   | 6291456   | 3145728   | 757301    | 1557.70  
Vertices per second, min: 707575, max: 1034798, avg: 884666
```

Roughly 40-60% faster compared to the master branch, depending on which computer I benchmark it on.

There's one big time sink left and that's the hash map for edges in `mesh::Mesh`. I didn't feel comfortable touching that since I'm not yet familiar enough with the `meshlite` codebase outside of `subdivide.rs`. I'm not going to touch that optimization, at least for a while, since I need to get back to actually using the library now :slightly_smiling_face: