@huxingyi 2018-10-09
Looks good, will merge it later when the CI done the check.
Currently, the bmesh.rs is heavily been used by Dust3D, so does mesh.rs. I have been rush for coding new features and don't have time to write tests for the mesh generating algorithm, but please feel free to change it if you think needed.
Currently, the speed of mesh generating algorithm of bmesh is ok with Dust3D, the bottleneck is the mesh union algorithm of CGAL, which is quite slow. however, the mesh union algorithm in the mesh.rs is very simple and can't handle complex mesh.