@anderejd 2018-10-10
> but please feel free to change it if you think needed.
Cool 👍  I'm quite tempted to try so it might happen later this week or next week.